doing user lab permissions
