+++ src/controllers/lab/lab_controllers.js
@@ -10,169 +10,7 @@ const userModels = require("../../models/user/user_models");
 const instituteModels = require("../../models/institute/institute_models");
 const jwt = require("jsonwebtoken");
 
-// O========================================================================================O
 
-/*
-    O=================================================================O
-    |   Funções de control relacionadas a inserção de laboratórios    |
-    O=================================================================O
-
-    Funções relacionadas a inserção de laboratórios:
-    - [X] CreateLab;
-    - [X] CreateLabUser;
-*/
-
-// O========================================================================================O
-
-// Função para criar um laboratório:
-const CreateLab = async (req, res) => {
-    const { sala, capacity } = req.body;
-
-    // Verificando se o usuário tem permissão para criar um laboratório:
-    const token = req.headers["x-access-token"];
-    const userID = jwt.decode(token).userID;
-
-    const getUserByID = await userModels.getUserByID(userID);
-
-    if (getUserByID.status === false) {
-        return res.status(400).json({
-            status: false,
-            message: "Usuário não encontrado!"
-        });
-    }
-
-    if (getUserByID.userData.CampusAdminLevel === 1 || getUserByID.userData.tipo === 1) {
-        return res.status(400).json({
-            status: false,
-            message: "Usuário não tem permissão para criar o laboratório!"
-        });
-    }
-
-    // Verificando se o campus existe:
-    const GetCampusByID = await instituteModels.getCampusByID(getUserByID.userData.ID_campus);
-
-    if (GetCampusByID.status === false) {
-        return res.status(400).json({
-            status: false,
-            message: "Campus não encontrado!"
-        });
-    }
-
-    // Verificando se o laboratório já existe:
-    const GetLabByName = await labModels.GetLabByName(sala, getUserByID.userData.ID_campus);
-
-    if (GetLabByName.status === true) {
-        return res.status(400).json({
-            status: false,
-            message: "Laboratório já existe!"
-        });
-    }
-
-    // Criando o laboratório:
-    const CreateLab = await labModels.CreateLab(sala, capacity, getUserByID.userData.ID_campus);
-
-    if (CreateLab.status === false) {
-        return res.status(400).json({
-            status: false,
-            message: "Erro ao criar laboratório!"
-        });
-    }
-
-    // Relacionando o laboratório com o usuário criador:
-    const CreateLabUser = await labModels.AddUser(CreateLab.labID, userID, 3);
-
-    if (CreateLabUser.status === false) {
-        return res.status(400).json({
-            status: false,
-            message: "Erro ao relacionar usuário com laboratório!"
-        });
-    }
-
-    return res.status(200).json({
-        status: true,
-        message: "Laboratório criado com sucesso!"
-    });
-};
-
-const CreateLabUser = async (req, res) => {
-    const { lab_id, user_id } = req.body;
-
-    // Verificando se o usuário existe:
-    const token = req.headers["x-access-token"];
-    const userID = jwt.decode(token).userID;
-
-    const GetUserByID = await userModels.getUserByID(userID);
-
-    if (GetUserByID.status === false) {
-        return res.status(400).json({
-            status: false,
-            message: "Usuário não encontrado!"
-        });
-    }
-
-    // Verificando se o laboratório existe:
-    const GetLabByID = await labModels.GetLabById(lab_id);
-
-    if (GetLabByID.status === false) {
-        return res.status(400).json({
-            status: false,
-            message: "Laboratório não encontrado!"
-        });
-    }
-
-    // Verificando permissão do usuário:
-    const GetLabUser = await labModels.GetLabUser(lab_id, userID);
-
-    if (GetLabUser.status === false || GetLabUser.data.AdminLevel !== 3 || GetUserByID.userData.ID_campus !== GetLabByID.data.ID_campus) {
-        return res.status(400).json({
-            status: false,
-            message: "Usuário não tem permissão para adicionar um usuário!"
-        });
-    }
-
-    // Verificando se o usuário existe:
-    const GetNewUserData = await userModels.getUserByID(user_id);
-
-    if (GetNewUserData.status === false) {
-        return res.status(400).json({
-            status: false,
-            message: "Usuário não encontrado!"
-        });
-    }
-
-    // Verificando se o usuário já pertence ao laboratório:
-    const GetNewUserRelation = await labModels.GetLabUser(lab_id, user_id);
-
-    if (GetNewUserRelation.status === true) {
-        return res.status(400).json({
-            status: false,
-            message: "Usuário já pertence ao laboratório!"
-        });
-    }
-
-    // Verificando se o campus do usuário é o mesmo do laboratório:
-    if (GetNewUserData.userData.ID_campus !== GetLabByID.data.ID_campus) {
-        return res.status(400).json({
-            status: false,
-            message: "Usuário não pertence ao mesmo campus do laboratório!"
-        });
-    }
-
-    // Adicionando o usuário ao laboratório:
-    const AddUser = await labModels.AddUser(lab_id, user_id, 1);
-
-    if (AddUser.status === false) {
-        return res.status(400).json({
-            status: false,
-            message: "Erro ao adicionar usuário!"
-        });
-    }
-
-    return res.status(200).json({
-        status: true,
-        message: "Usuário adicionado com sucesso!"
-    });
-};
 
 // O========================================================================================O
